Streamlit/requirements.txt
kmodes
streamlit
scikit-learn
pandas
numpy
pillow
datetime
plotly